% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterme.R
\name{filterme}
\alias{filterme}
\title{Filters a df}
\usage{
filterme(data, columns, filter_by_column, filters)
}
\arguments{
\item{data}{A df, like from the output of organizeme()}

\item{columns}{The columns one wants to select to plot in a graph}

\item{filter_by_column}{The column you want to filter rows based on}

\item{filters}{The rows you want to filter from said column}
}
\value{
A df to plot in a graph
}
\description{
Filters a df
}
