% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeme.R
\name{organizeme}
\alias{organizeme}
\title{Organize output from minmax()}
\usage{
organizeme(data, minmax_output)
}
\arguments{
\item{data}{The original data set with the column that has been normalized}

\item{minmax_output}{The output of the minmax() function}
}
\value{
A new df with an additional column containing the normalized values
}
\description{
Organize output from minmax()
}
