% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotme.R
\name{plotme}
\alias{plotme}
\title{Plots a Linegraph}
\usage{
plotme(data, x_axis, y_axis, Title)
}
\arguments{
\item{data}{A df, like the filterme() output}

\item{x_axis}{the column that represents the x axis on the linegraph}

\item{y_axis}{the column that represents the y axis on the linegraph}

\item{Title}{The title the graph needs to have}
}
\value{
A ggplot graph
}
\description{
Plots a Linegraph
}
